Fix removing all files while editing a message not working after having removed and re-added the same file Fix a bug where if the user edited a message with a file attached to it by performing these steps in the message editor:   - Remove the file   - Re-add the file   - Removes all files   - Send the edited message to the backend
The original file would continue to appear attached to the message after being edited, and if the user reloaded the page, the file would appear twice in the message. The file would also be sent to the backend after pressing the "Send" button, which means it would get saved to the database as a new file, and the message would end up having two files, which were the same, linked to it.